Added the serde_support feature
This change adds the serde_support feature, which when enabled adds
 Serialize and Deserialize implementations for

* KeyPair,
* SecretKey,
* PublicKey, and
* Signature

The Serialization is done manual, because of the of the large size of the bytes array in SecretKey
(64) and Signature (64) and for consistency also for PublicKey.
All these Types are serialized to a NeytypeStruct holding a bytes array.
While the NewtypeStruct should be ignored by the Serializer implementations

> Serializers are encouraged to treat newtype structs as insignificant wrappers
> around the data they contain. A reasonable implementation would be to forward
> to value.serialize(self).

But not all do.

Moved all serde related parts into a separate module.